+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ### Added
+- Endpoint to clear Local AAC Cache
 
 ### Changed
 - Updated Inference Sensemaker to only accept Observation Audit Log Events